More tidy up. Separated concept of Representation (Compute) and actual converts.
