+++ Speckle_Engine/Convert/ToBHoM/Geometry/Point.cs
@@ -32,6 +32,13 @@ namespace BH.Engine.Speckle
 {
     public static partial class Convert
     {
+        // -------------------------------------------------------------------------------- //
+        // NOTE
+        // These ToBHoM methods are not automatically called by any method in the Toolkit,
+        // as the deserialisation already brings back the BHoM object.
+        // Kept for reference and for manual use in the UI.
+        // -------------------------------------------------------------------------------- //
+
         [Description("Convert Speckle Point to BHoM Point")]
         public static BHG.Point ToBHoM(this SCG.SpecklePoint specklePoint)
         {
