+++ Speckle_Engine/Speckle_Engine.csproj
@@ -148,9 +148,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Compute\MeshRepresentation\Bar.cs" />
     <Compile Include="Convert\AdapterId.cs" />
-    <Compile Include="Convert\FromBHoM\BHoMObject.cs" />
-    <Compile Include="Convert\FromBHoM\Geometry\IGeometry.cs" />
+    <Compile Include="Convert\FromBHoM\Other\BHoMObject.cs" />
+    <Compile Include="Convert\FromBHoM\Geometry\_IGeometry.cs" />
     <Compile Include="Convert\ToBHoM\Geometry\Mesh.cs" />
     <Compile Include="Convert\ToBHoM\Geometry\Line.cs" />
     <Compile Include="Convert\ToBHoM\Geometry\Vector.cs" />
@@ -158,14 +159,13 @@
     <Compile Include="Convert\_GeomHelpers.cs" />
     <Compile Include="Convert\FromBHoM\Geometry\Mesh.cs" />
     <Compile Include="Convert\FromBHoM\Geometry\Line.cs" />
-    <Compile Include="Convert\FromBHoM\Structure\Bar.cs" />
-    <Compile Include="Convert\FromBHoM\Structure\Node.cs" />
-    <Compile Include="Convert\FromBHoM\Request.cs" />
+    <Compile Include="Convert\FromBHoM\SpeckleRepresentation\Bar.cs" />
+    <Compile Include="Convert\FromBHoM\SpeckleRepresentation\Node.cs" />
+    <Compile Include="Convert\FromBHoM\Other\Request.cs" />
     <Compile Include="Convert\FromBHoM\Geometry\Vector.cs" />
     <Compile Include="Convert\FromBHoM\Geometry\Point.cs" />
     <Compile Include="Convert\FromBHoM\_IFromBHoM.cs" />
     <Compile Include="Convert\ToBHoM\ToBHoM.cs" />
-    <Compile Include="Convert\ToBHoM\IToBHoM.cs" />
     <Compile Include="Compute\MeshRepresentation\Node.cs" />
     <Compile Include="Query\FilterBySpeckleGUID.cs" />
     <Compile Include="Query\FilterByBHoMGUID.cs" />
@@ -176,7 +176,6 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="Convert\ToBHoM\Structure\" />
     <Folder Include="Modify\" />
   </ItemGroup>
   <ItemGroup>
